hla tab on phenotype table
