+++ ui/src/components/HLA/HLA.tsx
@@ -11,24 +11,10 @@ import Search from "./HLASearch";
 import defaultConfig from "./HLAConfig";
 import './style.css';
 import ReactTooltip from "react-tooltip";
+import HLATable from "./HLATable";
 
 declare let window: ConfigurationWindow;
 const config: { [key: string]: any } = window?.config?.userInterface?.coding?.config || defaultConfig;
-
-const tableColumns: Column<HLAModel.Row>[] = hlaTableColumns as Column<HLAModel.Row>[]
-
-const tableProperties = {
-  defaultPageSize: 30
-}
-
-const defaultSorted = [{
-  id: 'phenocode',
-  desc: false
-},
-{
-  id: 'mlogp',
-  desc: true
-}]
 
 export const hasError = (errorMessage: string | null | undefined, content: JSX.Element): JSX.Element => {
   if (errorMessage === null || errorMessage === undefined) {
@@ -57,14 +43,6 @@ const HLA = (props) => {
     }
   }, []);
 
-  const prop: DownloadTableProps<HLAModel.Data, HLAModel.Row> = {
-    filename: 'hla.tsv',
-    tableData: hlaData,
-    dataToTableRows: (data) => data,
-    tableColumns,
-    tableProperties,
-    defaultSorted,
-  }
   const content = (
     <div className="hla">
       <div>
@@ -74,7 +52,7 @@ const HLA = (props) => {
       </div>
       <Search />
       <br/>
-      <CommonDownloadTable {...prop} />
+      <HLATable {...{data: hlaData}}/>
     </div>
   )
   return hlaData == null && props.error == null ? commonLoading : hasError(props.error, content)
