+++ ui/src/components/Phenotype/PhenotypeContext.tsx
@@ -1,11 +1,12 @@
 import React, { createContext, useEffect, useState } from "react";
 import { CredibleSet, PhenotypeParams, PhenotypeVariantData, QQ } from "./phenotypeModel";
 import { Phenotype } from "./../../common/commonModel";
-import { getUKBBN, getManhattan, getPhenotype, getCredibleSets, getQQ } from "./phenotypeAPI";
+import { getUKBBN, getManhattan, getPhenotype, getCredibleSets, getQQ, getHLA } from "./phenotypeAPI";
 import {setPageTitle} from "../../common/commonUtilities";
-
+import {HLAModel} from "../HLA/HLAModel";
 export interface PhenotypeState {
   phenotypeVariantData : PhenotypeVariantData
+  hlaData : HLAModel.Data
   UKBBN : Phenotype[]
   phenotype : Phenotype
   credibleSets : CredibleSet[]
@@ -34,13 +35,15 @@ const PhenotypeContextProvider = (props : Props) => {
   const [loading, setLoading] = useState<boolean| undefined>(true);
   const [qq, setQQ] = useState<QQ| undefined>(undefined)
   const [errorMessage, setErrorMessage] = useState<string| undefined>(undefined);
-
+  const [hlaData, setHlaData] = useState<HLAModel.Data| null>(null);
+  
   useEffect(() => {
     getManhattan(phenotypeCode,setPhenotypeVariantData);
     getUKBBN(phenotypeCode,setUKBBN);
     getPhenotype(phenotypeCode, setPhenotype, setErrorMessage);
     getCredibleSets(phenotypeCode, setCredibleSets);
     getQQ(phenotypeCode, setQQ);
+    getHLA(phenotypeCode, setHlaData);
   },[setPhenotypeVariantData, setUKBBN, setPhenotype, phenotypeCode]);
 
   useEffect(() => {
@@ -51,6 +54,7 @@ const PhenotypeContextProvider = (props : Props) => {
 <PhenotypeContext.Provider value={{
   phenotypeCode,
   phenotypeVariantData ,
+  hlaData ,
   UKBBN ,
   phenotype ,
   loading ,
