+++ ui/src/components/Phenotype/phenotypeAPI.tsx
@@ -3,6 +3,7 @@ import { compose, get, Handler } from "../../common/commonUtilities";
 import { CredibleSet, LocusGroupEntry, PhenotypeVariantData, QQ } from "./phenotypeModel";
 import { resolveURL } from "../Configuration/configurationModel";
 import { pValueSentinel } from "../../common/commonTableColumn";
+import { HLAModel } from "../HLA/HLAModel";
 
 const reshapeManhattan = (phenotypeCode: string) =>(data : PhenotypeVariantData) : PhenotypeVariantData => {
   if(data === null || data === undefined) return data;
@@ -79,4 +80,10 @@ export const getQQ = (phenotypeCode : string,
                       sink: (s: QQ) => void,
                       getURL = get) : void => {
   getURL(resolveURL(`/api/qq/pheno/${phenotypeCode}`), sink)
+}
+
+export const getHLA = (phenotypeCode : string,
+                      sink: (s: HLAModel.Data) => void,
+                      getURL = get) : void => {
+  getURL(resolveURL(`/api/v1/hla/phenocode/${phenotypeCode}`), sink)
 }
